test: organize integration tests following js integration tests
In order for developers to easily compare integration test coverage vs
the JS suite, a package with known well coverage, we organize the
integration tests to mirror those.

Specifically we chose filenames that are similar to those used in the
JS integration test repos. Then, in order to parallel the `describe`
blocks, we introduce a `mod` for each describe. The individual tests
therein should be parallel to the `it` calls in the JS repo.

This change also allows developers to easily run a subset of the
tests. A developer can run `cargo t <modname>` to just run the tests
therein. There is also IDE support for running all tests in a mod (see
below). Additionally, this allows us to shorten test function names
and use the same test function name for common
concerns ("nonexistent_cache").

In this PR we've organized the data plane commands into the following
files:
- cache_scalar.rs: commands where the value type is scalar
- cache_sorted_set.rs: commands where the value type is sorted set
- cache_key_exists.rs: commands to test existence of a cache key
- cache_item.rs: commands that apply to a cache item irrespective of
type.

Previously the tests under `cache_item.rs` had been grouped under
`cache_scalar.rs`. This wasn't as cohesive as the scalar tests group
commands where the value is a scalar type whereas the item tests apply
to all cache keys.

In the future we ought to group the cache tests under a folder `cache`
and rename the integration tests as appropriate.
